requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.23
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12
jinja2>=3.1.2
aiofiles>=24.1.0
pillow>=10.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
starlette>=0.37.0
pandas>=2.0.0
openpyxl>=3.1.0

